created authReducer module
